# backend/docs/openapi-pretty.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/expenses/{expenseId}" : {
      "put" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "updateExpense",
        "parameters" : [ {
          "name" : "expenseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "description",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "amount",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number"
          }
        }, {
          "name" : "category",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "FOOD_DRINKS", "TRANSPORTATION", "ACCOMMODATION", "ENTERTAINMENT", "SHOPPING", "UTILITIES", "HEALTHCARE", "EDUCATION", "TRAVEL", "GROCERIES", "RESTAURANTS", "GAS", "PARKING", "OTHER" ]
          }
        }, {
          "name" : "notes",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "updatedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "deleteExpense",
        "parameters" : [ {
          "name" : "expenseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "deletedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/expenses/group/{groupId}/share-split" : {
      "post" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "createExpenseWithShareSplits",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ShareSplitRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/group/{groupId}/percentage-split" : {
      "post" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "createExpenseWithPercentageSplits",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PercentageSplitRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/group/{groupId}/equal-split" : {
      "post" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "createExpenseWithEqualSplits",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateExpenseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/group/{groupId}/custom-split" : {
      "post" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "createExpenseWithCustomSplits",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomSplitRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/attachments/expense/{expenseId}" : {
      "get" : {
        "tags" : [ "attachment-controller" ],
        "operationId" : "getExpenseAttachments",
        "parameters" : [ {
          "name" : "expenseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AttachmentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "attachment-controller" ],
        "operationId" : "uploadFile",
        "parameters" : [ {
          "name" : "expenseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "uploadedByUserId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "description",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AttachmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/expenses/user/{userId}" : {
      "get" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "getUserExpenses",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Expense"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/user/{userId}/stats" : {
      "get" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "getUserExpenseStats",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseStats"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/group/{groupId}" : {
      "get" : {
        "tags" : [ "expense-controller" ],
        "operationId" : "getGroupExpenses",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Expense"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "profilePictureUrl" : {
            "type" : "string"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastLoginAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "Attachment" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "originalFilename" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "storedFilename" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "filePath" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "fileUrl" : {
            "type" : "string"
          },
          "fileSize" : {
            "type" : "integer",
            "format" : "int64"
          },
          "contentType" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "attachmentType" : {
            "type" : "string",
            "enum" : [ "RECEIPT_IMAGE", "DOCUMENT_IMAGE", "PDF_DOCUMENT", "INVOICE", "CONTRACT", "OTHER" ]
          },
          "description" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "deletedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "imageWidth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "imageHeight" : {
            "type" : "integer",
            "format" : "int32"
          },
          "fileChecksum" : {
            "type" : "string"
          },
          "storageProvider" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "fileExtension" : {
            "type" : "string"
          },
          "formattedFileSize" : {
            "type" : "string"
          },
          "image" : {
            "type" : "boolean"
          },
          "allowedFileType" : {
            "type" : "boolean"
          },
          "pdf" : {
            "type" : "boolean"
          },
          "thumbnailUrl" : {
            "type" : "string"
          },
          "deleted" : {
            "type" : "boolean"
          }
        },
        "required" : [ "contentType", "filePath", "originalFilename", "storedFilename" ]
      },
      "Expense" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 1
          },
          "amount" : {
            "type" : "number",
            "minimum" : 0.01
          },
          "date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "attachmentUrl" : {
            "type" : "string"
          },
          "attachmentFilename" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string",
            "enum" : [ "FOOD_DRINKS", "TRANSPORTATION", "ACCOMMODATION", "ENTERTAINMENT", "SHOPPING", "UTILITIES", "HEALTHCARE", "EDUCATION", "TRAVEL", "GROCERIES", "RESTAURANTS", "GAS", "PARKING", "OTHER" ]
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "DRAFT", "CONFIRMED", "CANCELLED", "DELETED" ]
          },
          "notes" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "formattedAmount" : {
            "type" : "string"
          },
          "fullySettled" : {
            "type" : "boolean"
          },
          "ageInDays" : {
            "type" : "integer",
            "format" : "int64"
          },
          "unsettledAmount" : {
            "type" : "number"
          },
          "splitAmountValid" : {
            "type" : "boolean"
          },
          "totalSplitAmount" : {
            "type" : "number"
          },
          "participantCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "activeAttachments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Attachment"
            },
            "uniqueItems" : true
          }
        },
        "required" : [ "amount", "date", "description" ]
      },
...
      "ShareSplitRequest" : {
        "type" : "object",
        "properties" : {
          "paidByUserId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number"
          },
          "userShareMap" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "category" : {
            "type" : "string"
          }
        }
      },
      "PercentageSplitRequest" : {
        "type" : "object",
        "properties" : {
          "paidByUserId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number"
          },
          "userPercentageMap" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "number"
            }
          },
          "category" : {
            "type" : "string"
          }
        }
      },
      "CreateExpenseRequest" : {
        "type" : "object",
        "properties" : {
          "paidByUserId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number"
          },
          "participantUserIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "category" : {
            "type" : "string"
          }
        }
      },
      "CustomSplitRequest" : {
        "type" : "object",
        "properties" : {
          "paidByUserId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number"
          },
          "userAmountMap" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "number"
            }
          },
          "category" : {
            "type" : "string"
          }
        }
      },
...
      "AttachmentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "expenseId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "uploadedBy" : {
            "$ref" : "#/components/schemas/UserResponse"
          },
          "originalFilename" : {
            "type" : "string"
          },
          "storedFilename" : {
            "type" : "string"
          },
          "fileSize" : {
            "type" : "integer",
            "format" : "int64"
          },
          "contentType" : {
            "type" : "string"
          },
          "attachmentType" : {
            "type" : "string",
            "enum" : [ "RECEIPT_IMAGE", "DOCUMENT_IMAGE", "PDF_DOCUMENT", "INVOICE", "CONTRACT", "OTHER" ]
          },
          "description" : {
            "type" : "string"
          },
          "checksum" : {
            "type" : "string"
          },
          "storageProvider" : {
            "type" : "string"
          },
          "thumbnailUrl" : {
            "type" : "string"
          },
          "isDeleted" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "deletedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ExpenseStats" : {
        "type" : "object",
        "properties" : {
          "totalExpenses" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalAmount" : {
            "type" : "number"
          },
          "settledAmount" : {
            "type" : "number"
          },
          "averageExpenseAmount" : {
            "type" : "number"
          },
          "unsettledAmount" : {
            "type" : "number"
          },
          "settlementPercentage" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
...
    }
  }
}